New version of the Occupations feature: v0.2
Changed the Views landing page to incorporate city & country into one column.
Added 'founded on:' field in order to sort by oldest occupation because sorting by title was meaningless since every node begins with 'Occupy ___' and that also begs for user abuse by adding special chars in order to rise to the top.
Changed the landing page URL from 'occupation' to 'occupations' to indicate the plurality of the view in deference to a page with a single occupation on it.
Changed the Occupation content type's image storage destination from 'images/occupation' to 'images/occupations/thumbnails' in order to allow for future image additions.

Signed-off-by: Senpai <joel@transparatech.com>